Add files via upload
🚢 Excited to share my latest project during Afame Technologies: Titanic Survival Prediction using Power BI! 🚢

I'm thrilled to unveil my latest endeavor where I utilized Power BI to delve into the fascinating world of data analysis and prediction. In this project, I explored the iconic Titanic dataset to predict survival outcomes based on various factors such as age, gender, ticket class, and more.

Using Power BI's robust tools and visualization capabilities, I was able to uncover insightful patterns and trends within the data, ultimately constructing a predictive model to estimate the likelihood of survival for passengers aboard the Titanic.

This project was not only a fantastic opportunity to hone my skills in data analysis and visualization but also to delve into the intriguing history of the Titanic and its passengers.

Stay tuned for a detailed walkthrough of my methodology and findings! #DataAnalysis #PowerBI #TitanicPrediction #DataVisualization